+++ derive-input/src/api/dsl/column.rs
@@ -9,7 +9,7 @@ pub struct SpacetimeDSLColumn {
     pub is_option: bool,
     // Only `Some(T)` if it has `#[wrap(name = MyTableId)]` or `#[wrapped(path = path::to::MyTableId)]`.
     pub wrapper_type: Option<WrapperType>,
-    // Only `Some(T)` if it has `#[foreign_key(table = my_table, column = my_column, action = ForeignKeyAction)]``.
+    // Only `Some(T)` if it has `#[foreign_key(table = my_table, column = my_column, on_delete = OnDeleteStrategy)]`.
     pub foreign_key: Option<ForeignKey>,
     pub getter: Getter,
     // Only `Some(T)` if mutable
