+++ derive-input/src/api/dsl/getter.rs
@@ -1,4 +1,4 @@
-/// TODO: Doc comment
+/// TODO: Doc comment field
 #[cfg_attr(feature = "clone", derive(Clone))]
 #[cfg_attr(feature = "debug", derive(Debug))]
 #[cfg_attr(feature = "partial-eq", derive(PartialEq))]
