+++ derive-input/src/api/dsl/setter.rs
@@ -1,6 +1,6 @@
 use crate::api::rust::RustVisibility;
 
-/// TODO: Doc comment
+/// TODO: Doc comment field
 #[cfg_attr(feature = "clone", derive(Clone))]
 #[cfg_attr(feature = "debug", derive(Debug))]
 #[cfg_attr(feature = "partial-eq", derive(PartialEq))]
