+++ derive-input/src/api/dsl/table.rs
@@ -5,6 +5,7 @@
 #[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
 pub struct SpacetimeDSLTable {
     pub plural_name: Box<str>,
+    // TODO: Implement
     pub on_delete_hooks: Vec<OnDeleteHook>,
     pub is_mutable: bool,
     pub has_created_at_column: bool,
